Completed refactoring code into object oriented paradigm.
As a result, removed the .kv file, as it detracts from this paradigm.

The structure of the code is now that of a tree, with the root node the screen manager. This creates the first screen (main menu) by instantiating the a subclass of screen manager, passing itself as an argument. This instance of screen_manager instantiates other screens from screen subclasses in the same way. The screens then instantiate subclasses of other widgets. The original screen manager is passed as an argument all the way down, as each new screen that's created needs to be added to it as a widget.

Added wrappers that can be used as decorators to cause transitions to slide left or right.

Added the Computer Modern fonts, which are now used for the orbifold names of wallpaper groups.